requirements.txt
beautifulsoup4
lxml
requests
pandas
python-telegram-bot
gspread
oauth2client
python-dotenv
psycopg2-binary
